Add build check for commits on a push
Functionality added: can now check whether commits pushed in a push event build or not.
This is done by parsing the POST request for commits. The JSON array of commits are handled by the function buildCommits that return an ArrayList of boolean values.
This ArrayList represents whether a build successfully builds or not. The index corresponds to the commit in the POST request.