Fix constructLog linking issues
This fixes so that construct log now links everything correctly. It does this by the user specifieng which directory that build logs should be hosted in. This adds the index.html to that folder and in turn create a subfolder with the builds logs in. Making everything easy to link together.